+++ api/src/Controller/Api/V1/ProductController.php
@@ -12,6 +12,7 @@ use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 /**
  * @Route("/product", name="api_product_")
@@ -21,21 +22,23 @@ class ProductController extends AbstractController
     private $serializer;
     private $repo;
     private $catRepo;
+    private $validator;
 
-    public function __construct(SerializerInterface $serializer, ProductRepository $repo, CategoryRepository $catRepo)
+    public function __construct(SerializerInterface $serializer, ProductRepository $repo, CategoryRepository $catRepo, ValidatorInterface $validator)
     {
         $this->serializer = $serializer;
         $this->repo = $repo;
         $this->catRepo = $catRepo;
+        $this->validator = $validator;
     }
 
     /**
      * @Route("/featured", methods={"GET"}, name="featured")
      */
-    public function featuredProducts(Request $request, array $currencies): JsonResponse
+    public function featuredProducts(Request $request, array $currencies): Response
     {
         $products = null;
-        $currency = $request->query->get('currency', null); // @TODO: Validate this parameter
+        $currency = $request->query->get('currency', null);
         if ($currency == null) {
             return new JsonResponse($this->repo->getFeaturedProducts());
         } elseif (in_array($currency, $currencies)) {
@@ -55,14 +58,24 @@ class ProductController extends AbstractController
     /**
      * @Route("/", methods={"POST"}, name="post")
      */
-    public function postProduct(Request $request): JsonResponse
+    public function postProduct(Request $request): Response
     {
-        $product = $this->serializer->deserialize($request->getContent(), Product::class, 'json', [AbstractNormalizer::IGNORED_ATTRIBUTES => ['category']]);
+        try {
+            $product = $this->serializer->deserialize($request->getContent(), Product::class, 'json', [AbstractNormalizer::IGNORED_ATTRIBUTES => ['category']]);
+        } catch (\UnexpectedValueException $e) {
+            return new Response($e->getMessage(), Response::HTTP_BAD_REQUEST);
+        }
+        //validate the product
+        $errors = $this->validator->validate($product);
+        if (count($errors) > 0) {
+            return new Response((string) $errors, Response::HTTP_BAD_REQUEST);
+        }
 
         $jsonObject = json_decode($request->getContent(), true);
+        //Check category
         if (isset($jsonObject['category']) && isset($jsonObject['category']['id'])) {
             $category = $this->catRepo->find($jsonObject['category']['id']);
-            // @TODO: Validate existence of category
+            if ($category == null) return new JsonResponse(['error' => 'Category not found'], Response::HTTP_BAD_REQUEST);
             $product->setCategory($category);
         }
 
@@ -73,7 +86,7 @@ class ProductController extends AbstractController
     /**
      * @Route("/{id}", methods={"GET"}, name="get_id")
      */
-    public function getProduct(?Product $product): JsonResponse
+    public function getProduct(?Product $product): Response
     {
         return $this->productToJsonResponse($product);
     }
@@ -81,16 +94,26 @@ class ProductController extends AbstractController
     /**
      * @Route("/{id}", methods={"PUT"}, name="put_id")
      */
-    public function putProduct(Request $request, ?Product $product): JsonResponse
+    public function putProduct(Request $request, ?Product $product): Response
     {
         if ($product != null) {
-            $this->serializer->deserialize($request->getContent(), Product::class, 'json', [AbstractNormalizer::OBJECT_TO_POPULATE => $product, AbstractNormalizer::IGNORED_ATTRIBUTES => ['category']]);
+            try {
+                $this->serializer->deserialize($request->getContent(), Product::class, 'json', [AbstractNormalizer::OBJECT_TO_POPULATE => $product, AbstractNormalizer::IGNORED_ATTRIBUTES => ['category']]);
+            } catch (\UnexpectedValueException $e) {
+                return new Response($e->getMessage(), Response::HTTP_BAD_REQUEST);
+            }
+            //validate the product
+            $errors = $this->validator->validate($product);
+            if (count($errors) > 0) {
+                return new Response((string) $errors, Response::HTTP_BAD_REQUEST);
+            }
+
 
             $jsonObject = json_decode($request->getContent(), true);
             if (isset($jsonObject['category'])) {
                 if (isset($jsonObject['category']['id'])) {
                     $category = $this->catRepo->find($jsonObject['category']['id']);
-                    // @TODO: Validate existence of category
+                    if ($category == null) return new Response('Category not found', Response::HTTP_BAD_REQUEST);
                     $product->setCategory($category);
                 } else {
                     $product->setCategory(null);
@@ -106,15 +129,21 @@ class ProductController extends AbstractController
     /**
      * @Route("/{id}", methods={"DELETE"}, name="delete_id")
      */
-    public function deleteProduct(?Product $product): JsonResponse
+    public function deleteProduct(?Product $product): Response
     {
+        if ($product == null) {
+            return $this->productToJsonResponse(null); // Return standard not found response
+        }
+
         $this->repo->remove($product);
         return new JsonResponse(['deleted' => true]);
     }
 
-    private function productToJsonResponse(null|array|Product $data, array $groups = ['rest']): JsonResponse
+    private function productToJsonResponse(null|array|Product $data, array $groups = ['rest']): Response
     {
-        if ($data == null) return new JsonResponse([], Response::HTTP_NO_CONTENT);
+        if ($data == null) {
+            return new Response('Not found', Response::HTTP_NO_CONTENT);
+        }
 
         return JsonResponse::fromJsonString(
             $this->serializer->serialize(
